fix avatar upload bug./ BUG: inconsistant avatar in userInfo and subject

+++ routes/users.js
@@ -12,26 +12,29 @@ router.get("/users/:id", function(req, res){
 
 router.post("/users/:id/avatar", middleware.isCurUser, function(req, res){
     var clientId = "8ae2612bf7cb691";
+    var oldDeleteHash = req.user.avatarDeleteHash;
 
-    request({
-        // delete the old avatar from imgur
-        method: "DELETE",
-        url: "https://api.imgur.com/3/image/" + req.user.avatarDeleteHash,
-        headers: {
-            "Authorization": "Client-ID " + clientId
-        }
-    }, function(err, ress, body){
-        console.log(err);
-        console.log(body);
+    // change the url and deleteHash in local database
+    req.user.avatarUrl = req.body.avatarUrl;
+    req.user.avatarDeleteHash = req.body.avatarDeleteHash;
+    req.user.save();
 
-        // change the url and deleteHash in local database
-        req.user.avatarUrl = req.body.avatarUrl;
-        req.user.avatarDeleteHash = req.body.avatarDeleteHash;
-        req.user.save();
+    if(oldDeleteHash !== ""){ // if the old avatar is not default avatar
+        request({
+            // delete the old avatar from imgur
+            method: "DELETE",
+            url: "https://api.imgur.com/3/image/" + oldDeleteHash,
+            headers: {
+                "Authorization": "Client-ID " + clientId
+            }
+        }, function(err, ress, body){
+            console.log(err);
+            console.log(body);
+            res.redirect("/users/" + req.params.id);
+        });
+    }else{
         res.redirect("/users/" + req.params.id);
-    });
-
-    
+    }
 });
 
 module.exports = router;
